Add minimal package structure, unit test, and tutorial
The package now contains a minimal structure (it can be imported and a
test function can be called), and a unit test that can be run with
nosetests. A tutorial.ipynb file has also been added, demonstrating how
the unit tests can be run and how the package can be used.